Only pull videos from results of YouTube search
Code had a prior bug where we'd get a 'KeyError' for 'videoId'. A fix
was proposed in commit 66391c2, however this would create strange
behavior for non-video resources that the API search would return.
Modified code to only parse videos in the results, as that is what the
code currently supports.

See
http://googleapis.github.io/google-api-python-client/docs/dyn/youtube_v3.search.html