Transform media files to static files.
Up until now, we have been serving our images as media files. Each image
was pulled in our templates using a dev-only URL of
'127.0.0.1:8000/media'. I've moved the media files as static files so
that we can serve them in prod using the 'whitenoise' module and use the
`{% static ... %}` syntax in our templates.